working integration with frequency switching
