+++ TagV3.0_U575VGT/Core/Inc/Recovery Inc/VHF.h
@@ -18,6 +18,8 @@
 //Defines
 #define TX_FREQ "144.3900"
 #define RX_FREQ "144.3900"
+#define DOMINICA_TX_FREQ "145.0500"
+#define DOMINICA_RX_FREQ "145.0500"
 #define VHF_VOLUME_LEVEL 4
 
 //Lengths of messages used to configure the VHF module
@@ -48,7 +50,7 @@
  	 -huart - UART handler to talk to the module
  	 -isHigh: whether or not to use high power (1W) or low power (0.5W). Passing true means high power.
 */
-HAL_StatusTypeDef initialize_vhf(UART_HandleTypeDef huart, bool isHigh);
+HAL_StatusTypeDef initialize_vhf(UART_HandleTypeDef huart, bool isHigh, char * txFreq, char * rxFreq);
 
 /*Configure the VHF module over UART
  Parameters:
@@ -58,7 +60,7 @@ HAL_StatusTypeDef initialize_vhf(UART_HandleTypeDef huart, bool isHigh);
  	 -lpf: Apply a low-pass-filter on the signals.
  	 -hps: Apply a high-pass-filter on the signals.
 */
-HAL_StatusTypeDef configure_dra818v(UART_HandleTypeDef huart, bool emphasis, bool lpf, bool hpf);
+HAL_StatusTypeDef configure_dra818v(UART_HandleTypeDef huart, bool emphasis, bool lpf, bool hpf, char * txFreq, char * rxFreq);
 
 //Enables or disbles the push-to-talk pin. Setting true lets us talk to the module (i.e., transmit signals).
 void set_ptt(bool isTx);
