+++ TagV3.0_U575VGT/Core/Src/Recovery Src/Aprs.c
@@ -6,209 +6,105 @@
  */
 
 #include "Recovery Inc/Aprs.h"
+#include "Recovery Inc/VHF.h"
 #include "Recovery Inc/GPS.h"
+#include "Recovery Inc/AprsPacket.h"
+#include "Recovery Inc/AprsTransmit.h"
 #include "main.h"
-#include <stdint.h>
-#include <string.h>
-#include <stdbool.h>
-#include <stdio.h>
 
-//Extern variables for HAL handlers
+//Extern variables for HAL UART handlers
 extern UART_HandleTypeDef huart4;
 extern UART_HandleTypeDef huart3;
 
-static void append_flag(uint8_t * buffer, uint8_t numFlags);
-static void append_callsign(uint8_t * buffer, char * callsign, uint8_t ssid);
-static void append_gps_data(uint8_t * buffer, float lat, float lon);
-static void append_other_data(uint8_t * buffer, uint16_t course, uint16_t speed, char * comment);
-static void append_frame_check(uint8_t * buffer, uint8_t buffer_length);
+static bool toggleFreq(bool is_gps_dominica, bool is_currently_dominica);
 
 void aprs_thread_entry(ULONG aprs_thread_input){
+
+	//buffer for packet data
+	uint8_t packetBuffer[APRS_PACKET_MAX_LENGTH] = {0};
 
 	//Create the GPS handler and configure it
 	GPS_HandleTypeDef gps;
 	initialize_gps(&huart3, &gps);
 
 	//Initialize VHF module for transmission. Turn transmission off so we don't hog the frequency
-	HAL_StatusTypeDef vhf_init_ret = initialize_vhf(huart4, false);
+	initialize_vhf(huart4, false, TX_FREQ, RX_FREQ);
 	set_ptt(false);
 
+	//We arent in dominica by default
+	bool isInDominica = false;
+
+	//Main task loop
 	while(1){
 
-	}
-}
+		//GPS data struct
+		GPS_Data gps_data;
 
-void aprs_generate_packet(float lat, float lon){
+		//Attempt to get a GPS lock
+		bool isLocked = get_gps_lock(&gps, &gps_data);
 
-	uint8_t buffer[255] = {0};
-	append_flag(buffer, 150);
+		//The time we will eventually put this task to sleep for. We assign this assuming the GPS lock has failed (only sleep for a shorter, fixed period of time).
+		//If we did get a GPS lock, the sleepPeriod will correct itself by the end of the task (be appropriately assigned after APRS transmission)
+		uint32_t sleepPeriod = GPS_SLEEP_LENGTH;
 
-	append_callsign(&buffer[150], APRS_DESTINATION_CALLSIGN, APRS_DESTINATION_SSID);
+		//If we've locked onto a position, we can start creating an APRS packet.
+		if (isLocked){
 
-	append_callsign(&buffer[157], APRS_SOURCE_CALLSIGN, APRS_SOURCE_SSID);
+			aprs_generate_packet(packetBuffer, gps_data.latitude, gps_data.longitude);
 
-	//We can also treat the digipeter as a callsign since it has the same format
-	append_callsign(&buffer[164], APRS_DIGI_PATH, APRS_DIGI_SSID);
-	buffer[170] += 1;
+			//We first initialized the VHF module with our default frequencies. If we are in Dominica, re-initialize the VHF module to use the dominica frequencies.
+			//
+			//The function also handles switching back to the default frequency if we leave dominica
+			isInDominica = toggleFreq(gps_data.is_dominica, isInDominica);
 
-	//Add the control ID and protocol ID
-	buffer[171] = APRS_CONTROL_FIELD;
-	buffer[172] = APRS_PROTOCOL_ID;
+			//Start transmission
+			set_ptt(true);
 
-	//Attach the payload (including other control characters)
-	//TODO: Replace with real GPS data.
-	append_gps_data(&buffer[173], 42.3636, -71.1259);
+			//Now, transmit the signal through the VHF module. Transmit a few times just for safety.
+			for (uint8_t transmits = 0; transmits < 3; transmits++){
+				aprs_transmit_send_data(packetBuffer, APRS_PACKET_LENGTH);
+			}
 
-	//Attach other information (course, speed and the comment)
-	append_other_data(&buffer[193], 360, 0, APRS_COMMENT);
+			//end transmission
+			set_ptt(false);
 
-	append_frame_check(buffer, 219);
-
-	append_flag(&buffer[221], 3);
-
-	HAL_Delay(100);
-}
-
-//Appends the flag character (0x7E) to the buffer 'numFlags' times.
-static void append_flag(uint8_t * buffer, uint8_t numFlags){
-
-	//Add numFlags flag characters to the buffer
-	for (uint8_t index = 0; index < numFlags; index++){
-		buffer[index] = APRS_FLAG;
-	}
-}
-
-//Appends a callsign to the buffer with its SSID.
-static void append_callsign(uint8_t * buffer, char * callsign, uint8_t ssid){
-
-	//Determine the length of the callsign
-	uint8_t length  = strlen(callsign);
-
-	//Append the callsign to the buffer. Note that ASCII characters must be left shifted by 1 bit as per APRS101 standard
-	for (uint8_t index = 0; index < length; index++){
-		buffer[index] = (callsign[index] << 1);
-	}
-
-	//The callsign field must be atleast 6 characters long, so fill any missing spots with blanks
-	if (length < APRS_CALLSIGN_LENGTH){
-		for (uint8_t index = length; index < APRS_CALLSIGN_LENGTH; index++){
-			//We still need to shift left by 1 bit
-			buffer[index] = (' ' << 1);
+			//Set the sleep period for a successful APRS transmission
+			sleepPeriod = APRS_BASE_SLEEP_LENGTH;
 		}
-	}
 
-	//Now, we've filled the first 6 bytes with the callsign (index 0-5), so the SSID must be in the 6th index.
-	//We can find its ASCII character by adding the integer value to the ascii value of '0'. Still need to shift left by 1 bit.
-	buffer[APRS_CALLSIGN_LENGTH] = (ssid + '0') << 1;
+		//Go to sleep now
+		tx_thread_sleep(sleepPeriod);
+	}
 }
 
-//Appends the GPS data (latitude and longitude) to the buffer
-static void append_gps_data(uint8_t * buffer, float lat, float lon){
+/*Reconfigures the VHF module to change the transmission frequency based on where the GPS is.
+ *
+ * is_gps_dominica: whether or not the current gps data is in dominica
+ * is_currently_dominica: whether or not our VHF module is configured to the dominica frequency
+ *
+ * Returns: the current state of our configuration (whether or not VHF is configured for dominica
+ */
+static bool toggleFreq(bool is_gps_dominica, bool is_currently_dominica){
 
-	//indicate start of real-time transmission
-	buffer[0] = APRS_DT_POS_CHARACTER;
+	//If the GPS is in dominica, but we are not configured for it, switch to dominica
+	if (is_gps_dominica && !is_currently_dominica){
 
-	//First, create the string containing the latitude and longitude data, then save it into our buffer
-	bool isNorth = true;
+		//Re-initialize for dominica frequencies
+		initialize_vhf(huart4, false, DOMINICA_TX_FREQ, DOMINICA_RX_FREQ);
 
-	//If we have a negative value, then the location is in the southern hemisphere.
-	//Recognize this and then just use the magnitude of the latitude for future calculations.
-	if (lat < 0){
-		isNorth = false;
-		isNorth *= -1;
+		//Now configured for dominica, return to indicate that
+		return true;
+	}
+	//elseif, we are not in dominica, but we are configured for dominica. Switch back to the regular frequencies
+	else if (!is_gps_dominica && is_currently_dominica){
+
+		//Re-initialize for default frequencies
+		initialize_vhf(huart4, false, TX_FREQ, RX_FREQ);
+
+		//No longer on dominica freq, return to indicate that
+		return false;
 	}
 
-	//The coordinates we get from the GPS are in degrees and fractional degrees
-	//We need to extract the whole degrees from this, then the whole minutes and finally the fractional minutes
-
-	//The degrees are just the rounded-down integer
-	uint8_t latDegWhole = (uint8_t) lat;
-
-	//Find the remainder (fractional degrees) and multiply it by 60 to get the minutes (fractional and whole)
-	float latMinutes = (lat - latDegWhole) * 60;
-
-	//Whole number minutes is just the fractional component.
-	uint8_t latMinutesWhole = (uint8_t) latMinutes;
-
-	//Find the remainder (fractional component) and save it to two decimal points (multiply by 100 and cast to int)
-	uint8_t latMinutesFrac = (latMinutes - latMinutesWhole) * 100;
-
-	//Find our direction indicator (N for North of S for south)
-	char latDirection = (isNorth) ? 'N' : 'S';
-
-	//Create our string. We use the format ddmm.hh(N/S), where "d" is degrees, "m" is minutes and "h" is fractional minutes.
-	//Store this in our buffer.
-	snprintf(&buffer[1], APRS_LATITUDE_LENGTH, "%02d%02d.%02d%c", latDegWhole, latMinutesWhole, latMinutesFrac, latDirection);
-
-
-	//Right now we have the null-terminating character in the buffer "\0". Replace this with our latitude and longitude seperating symbol "1".
-	buffer[APRS_LATITUDE_LENGTH] = APRS_SYM_TABLE_CHAR;
-
-	//Now, repeat the process for longitude.
-	bool isEast = true;
-
-	//If its less than 0, remember it as West, and then take the magnitude
-	if (lon < 0){
-		isEast = false;
-		lon *= -1;
-	}
-
-	//Find whole number degrees
-	uint8_t lonDegWhole = (uint8_t) lon;
-
-	//Find remainder (fractional degrees), convert to minutes
-	float lonMinutes = (lon - lonDegWhole) * 60;
-
-	//Find whole number and fractional minutes. Take two decimal places for the fractional minutes, just like before
-	uint8_t lonMinutesWhole = (uint8_t) lonMinutes;
-	uint8_t lonMinutesFractional = (lonMinutes - lonMinutesWhole) * 100;
-
-	//Find direction character
-	char lonDirection = (isEast) ? 'E' : 'W';
-
-	//Store this in the buffer, in the format dddmm.hh(E/W)
-	snprintf(&buffer[APRS_LATITUDE_LENGTH + 1], APRS_LONGITUDE_LENGTH, "%03d%02d.%02d%c", lonDegWhole, lonMinutesWhole, lonMinutesFractional, lonDirection);
-
-	//Appending payload character indicating the APRS symbol (using boat symbol). Replace the null-terminating character with it.
-	buffer[APRS_LATITUDE_LENGTH + APRS_LONGITUDE_LENGTH] = APRS_SYM_CODE_CHAR;
-}
-
-//Appends other extra data (course, speed and the comment)
-static void append_other_data(uint8_t * buffer, uint16_t course, uint16_t speed, char * comment){
-
-	//Append the course and speed of the tag (course is the heading 0->360 degrees
-	uint8_t length = 8 + strlen(comment);
-	snprintf(buffer, length, "%03d/%03d%s", course, speed, comment);
-}
-
-//Calculates and appends the CRC frame checker. Follows the CRC-16 CCITT standard.
-static void append_frame_check(uint8_t * buffer, uint8_t buffer_length){
-
-	uint16_t crc = 0xFFFF;
-
-	//Loop through each *bit* in the buffer. Only start after the starting flags.
-	for (uint8_t index = 150; index < buffer_length; index++){
-
-		uint8_t byte = buffer[index];
-
-		for (uint8_t bit_index = 0; bit_index < 8; bit_index++){
-
-			bool bit = (byte >> bit_index) & 0x01;
-
-			//Bit magic for the CRC
-			unsigned short xorIn;
-			xorIn = crc ^ bit;
-
-			crc >>= 1;
-
-			if (xorIn & 0x01) crc ^= 0x8408;
-
-		}
-	}
-
-	uint8_t crcLo = (crc & 0xFF) ^ 0xFF;
-	uint8_t crcHi = (crc >> 8) ^ 0xFF;
-
-	buffer[buffer_length] = crcLo;
-	buffer[buffer_length + 1] = crcHi;
+	//else: do nothing
+	return is_currently_dominica;
 }
