+++ TagV3.0_U575VGT/Core/Src/app_threadx.c
@@ -28,6 +28,7 @@
 #include "main.h"
 #include "Recovery Inc/AprsTransmit.h"
 #include "Recovery Inc/VHF.h"
+#include "Recovery Inc/Aprs.h"
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
@@ -47,8 +48,8 @@
 
 /* Private variables ---------------------------------------------------------*/
 /* USER CODE BEGIN PV */
-TX_THREAD test_thread;
-uint8_t thread_stack[THREAD_STACK_SIZE];
+TX_THREAD aprs_thread;
+uint8_t aprs_thread_stack[THREAD_STACK_SIZE];
 extern UART_HandleTypeDef huart4;
 extern uint8_t transmits;
 /* USER CODE END PV */
@@ -68,13 +69,13 @@ UINT App_ThreadX_Init(VOID *memory_ptr)
 {
   UINT ret = TX_SUCCESS;
   /* USER CODE BEGIN App_ThreadX_MEM_POOL */
-  VOID * pointer = &thread_stack;
+  VOID * pointer = &aprs_thread_stack;
   // Allocate memory pool
   TX_BYTE_POOL* byte_pool = (TX_BYTE_POOL*) memory_ptr;
   ret = tx_byte_allocate(byte_pool, &pointer, THREAD_STACK_SIZE, TX_NO_WAIT);
   /* USER CODE END App_ThreadX_MEM_POOL */
   /* USER CODE BEGIN App_ThreadX_Init */
-  tx_thread_create(&test_thread, "Test_Thread", test_thread_entry, 0x1234, thread_stack, THREAD_STACK_SIZE, 3, 3, TX_NO_TIME_SLICE, TX_AUTO_START);
+  tx_thread_create(&aprs_thread, "APRS_Thread", aprs_thread_entry, 0x1234, aprs_thread_stack, THREAD_STACK_SIZE, 3, 3, TX_NO_TIME_SLICE, TX_AUTO_START);
   /* USER CODE END App_ThreadX_Init */
 
   return ret;
